requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
click==8.1.3
    # via pip-tools
django==4.0.4
    # via
    #   -r requirements.in
    #   django-address
django-address==0.2.8
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.2
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
